Improve rule and defrule macros
rule/defrule macros have specs.

rule/defrule accepts a new clause `:as` which allows binding of the
value successfully matched by the `:replace` clause pattern.

  (rule []
    :replace {1 2}
    :as m
    :with ~m*
    :where [m* (assoc m 3 4)])

Here `m` is bound to `{1 2}` and, thus, `m*` becomes `{1 2 3 4}`.

Substitution maps are now keyed by strings (variable names). This
relieves the problem of resolving and extending with splicing and
non-splincing variables differently. Resolution and extension of the
variables `~@xs` and `~xs` are identical.