Add support for pattern aliases
This patch allows patterns to specify labels for subpatterns. For
example, the pattern

[^{:as xs} [~x ~y ~z] ~xs]

gives the pattern [~x ~y ~z] the name xs and can be reused throughout
the pattern. This pattern would match the form

[[1 2 3] [1 2 3]]

Labels may not conflict or be cyclic.

The pattern

[^{:as a} [~c]
 ^{:as b} [~a]
 ^{:as c} [~b]]

is illegal because it cyclic. The subpattern named by a references c,
which references b, which references a.

The pattern

[^{:as a} [1 2]
 ^{:as a} [3 4]]

would be illegal because a cannot have different definitions.