Added error message when zero or more is improperly used.
Before this change is someone accidentally used zero-or-more improperly `(... ?x)` with no preceding value, there would be invalid macroexpansion. Now users will recieve a much more helpful error message.

There are equivilant problems with one-or-more. But I'd like to also handle the case where a user forgets the number on one-or-more `(1 .. ?x)`. Unfortunately this leaves one bit of ugly code that I'm not sure how to make better. So I will make that a separate commit/PR from this one.