Added error cases for rp+
I decided to add two separate error cases under rp+. The first is if
you leave off the prefix that you want to repeat N times `(..2
?x). The second case I've encountered is leaving off the N on
`..N`. In this case, I think the best course of action is to ask the
user if they mean `..N` or `...` as those are easy to typo between.

One unforuntate part of this change was the error handling
`min-length`. If a user leaves off the N, the min-length can't be
taken from the regex of the dots. Unforunately, min-length is called
before check so some change here is necessary. I'm not totally happy with the
solution I have so if you have a better idea, feel free to change it.