Ground set patterns fully express subset semantics
Previously ground set patterns such as

  #{#{:foo} #{:bar}}

were compiled to a clojure.set/subset? check. However, such a check
was inaccurate. If applied to the pattern above the contained sets
would be treated as literal values and not subset patterns themselves
resulting in conflicting semantics and false negative matches. This
patch addresses this problem.

Matching the above pattern requires utilizing most of the same
machinery involved in matching a set search pattern with one
exception: it only needs to determine if a solution exists. To handle
this specific situation I have added a new :find tree node the match
compiler.

The :find tree node has the same shape as the :search tree node e.g.

  [:find element-symbol elements]

but, of couse, emits slightly different code when compiled to
Clojure. Instead of producing a stream of results and picking the
first one, we loop over the search space elements stopping immediately
when we find a match. This allows us to avoid the overhead of
producing intermediate collections.

This new addition also paves the way for us to provide a new macro
which has functionality that lies "in between" match and search.