Wrap memory variable updates with :save/:load
Compilation of the new :find node made it possible for certain
patterns containing memory variables to return incorrect results. This
seemed to occur at the intersection of scans, maps, and logic
variables. For example given the data

  (def pets
    [{:name "George"
      :species "Parakeet"
      :age 3
      :owners ["Frege" "Peirce"]}
     {:name "Francis"
      :species "Dog"
      :age 8
      :owners ["De Morgan"]}
     {:name "Bob"
      :species "Goldfish"
      :age 1
      :owners ["Peirce"]}])

the find

  (r.match/find pets
    [_ ...
     {:name !names
      :owners [_ ... ?owner . _ ...]}
     .
     (or {:name !names
          :owners [_ ... ?owner . _ ...]}
         _)
     ...]
    [?owner !names])

would incorrectly return

  ["Peirce" ["George" "Bob" "Francis"]]

which is not an element of the result set produced by search with the
same inputs. The problem is that the second occurence of key/value
pattern

  :name !names

would match successfully and update !names but

  :owners [_ ... ?owner . _ ...]

would subsequently fail. Due to how find is compiled the !names state
could persist through this failure via the

  _

pattern.