pattern-op-dispatch can return fully qualified symbols
pattern-op-dispatch now attempts to fully qualify symbols appearing at
the head of a list by

1. resolving namespaces aliases, or
2. resolving the symbol with respect to *ns*.

This allows new syntax extensions to be defined in a way that has
parity with Clojure's def (though more verbosely) and resolution of
symbols to vars.

Beacuse this feature depends on *ns*, it is only partially supported
in ClojureScript - via Clojure macros. Invokations of parse outside of
macros will need to fully qualify symbols which reference syntax
extensions manually. One simple way to achieve this would be to define
a macro which derives the alias-mapping of *ns* but from symbol? to
qualified-symbol?,

  (defmacro ns-alias-map []
    '~(into {}
          (map (juxt key (comp ns-name val)))
          (ns-aliases *ns*)))

save a reference to it in ClojureScript,

  (def
    ^{:arglists '([sym])}
    resolve-alias (ns-alias-map))

and then apply the mapping as neccessary.