Add defsyntax macro
This patch allows programmers to extend meander's pattern syntax by
extending the parser.

  (defsyntax like [regex]
    `(~'pred
       (fn [x]
         (and (string? x)
              (re-matches? ~regex x)))))

defsyntax is like defn and defines an ordinary function that takes
arguments and returns a result as usual. However, it also defines
methods for pattern-op and expand-usr-op with the fully qualified
function name (which pattern-op-dispatch will expand). The pattern-op
definition verifies the correct number of arguments are passed and, if
not, throws an error. The expand-usr-op definiton handles passing the
arguments to the defined function and calling parse recursively on
it's result.