Safely rewrite shared bindings
Previously rewrite-shared-bindings had to be commented out because it
naively rewrote bindings in an unsafe way. This patch adds a new
function gensym* which behaves like clojure.core/gensym but attaches
additional meta data which can be later used to determine which
bindings are safe to rewrite. In some cases rewrite-shared-bindings
can dramatically reduce the size of code generated and eliminate
redundant computation.