Initial gamma release
This new version of Meander introduces new syntax.

Vector, Seq, and Map patterns now support :as style as patterns as
common in Clojure

  (and [!xs !ys ...] ?xs)
  (and (!xs !ys ...) ?xs)
  (and {:k1 ?v1} ?m)

Can now be more conveniently written as

  [!xs !ys ... :as ?xs]
  (!xs !ys ... :as ?xs)
  {:k1 ?v1 :as ?m}

semantically there is no differnce.

A new "containment" pattern matching operator is now available. This
operator allows pattern matching on arbitrarily nested values and,
optionally, allows the binding of an function which edits the context
where the match was found.

  (find {:student "Harry Hacker"
         :scores [99 84 93 12 76]}
    {:student ?student
     :scores ($ ?context (pred #(and (number? %) (< % 70)) ?grade))}
    {:student ?student
     :scores (?context 81)}) ;;
  ;; =>
  {:student "Harry Hacker"
   :scores [99 84 93 81 76]}
  ;; ----------------^^

In the above example we use the containment pattern to look for a
number that is less than 70. Note, that we do not wrap it in a vector
to match the value of :scores, this is on purpose for this
demonstration. Next, notice on the right size we apply the ?context
function which updates the position in :scores where we found the
match and returns the new updated scores.

If were feeling a little more confident we could achieve the same
thing with just

  (find {:student "Harry Hacker"
         :scores [99 84 93 12 76]}
    ($ ?context (pred #(and (number? %) (< % 70)) ?grade))
    (?context 81))
  ;; =>
  {:student "Harry Hacker"
   :scores [99 84 93 81 76]}
  ;; ----------------^^

because the search is arbitary.

Another new feature is "rest-map" matching. This allows map patterns
to use the familiar "&" character but with maps and pattern matching
to capture the rest of the map not specified. In other words, it's a
declarative dissoc operation.

  (match {:student "Betty Bitshifter"
          :scores [29 34 73 42 66]}
    {:student ?student & ?rest-map}
    ?rest-map)
  ;; =>
  {:scores [29 34 73 42 66]}

This can be useful in rewriting situations where you want to match
parts of a map but may or may not want to retain them.

On the substitution side, rest-maps represent the inverse: a
declarative assoc operation.

  ((rewrite
    {:name (re #"^([^ ]+) *([^ ]+)" [_ ?first ?last])
     :as ?student}
    {:first-name ?first
     :last-name ?last
     & ?student})
   {:name "Betty Bitshifter"
    :address {:street "10 Bit Road"
              :country "Potatoes"}})
  ;; =>
  {:name "Betty Bitshifter",
   :address {:street "10 Bit Road", :country "Potatoes"},
   :first-name "Betty",
   :last-name "Bitshifter"}

In this example we parse the first and last name out of name and
associate with original data.