Fixes substitution looping infinitely when using :rp*.
In compile-substitution :rst nodes were being compiled into :rp* nodes with :elements. From what I can tell :rp* nodes typically have a :cat and then :cat has :elements. This was getting *fixed* in the :rp* compilation, but that was messing things up for nodes that didn't come from :rst. So I changed :rst to compile to the standard rp* format. Then I fixed rp* and rp+. All the tests pass for me where as before the test running hung on the substitution namespace.

I also added in a throw when we can't find memory variables. There was a comment that this was needed and I figure this throw is better behavior than an infinite loop.