Add rest and as pattern for sets
Up until now it was not possible to capture the "rest" of a set nor
was it possible to capture a whole set as conveniently as :as in other
collection type patters. This patch exploits the Clojure meta reader
syntaxes ^symbol and ^:keyword to emulate & and :as for set patterns
with ^& and ^:as.

The pattern

  #{1 ^& ?rest-set}

applied to #{1 2 3} would bind ?rest-set to #{2 3}. Likewise the
pattern

  #{1 ^:as ?whole-set}

applied to the same set would bind ?whole-set to #{1 2 3}.