Fix several bugs related to :wth and :map compilation
1. The derivation of the ref-spec-map was wrong. This patch correctly
computes all of the possible function signatures given a ref-map
2. For :map and :set patterns which use &, dissoc and disj were
respectively broken when used recursively from with. This is because
the logic variables generated by parse-expand end up being checked for
equality on recursive calls which, clearly, won't work. To fix this,
I've added a :mutable? field to :lvr nodes in parse-expand which
always get compiled to a :bind IR node. This is a hack and, probably,
the implies the need for a new type of AST node. Perhaps :any could
occupy that role?