Fix/Improve literal compilation
By converting literal? :seq or :vec nodes to :lit in parse-expand we
lose out on our respective seq? and vector? checks resulting in
inaccurately compiled code. A rule that is written for seq forms
should not match forms that are vectors accidentally, etc.

To fix the problem I moved the check for literal? inside the matrix
compilation methods for :seq and :vec nodes. Additionally, I changed
how code is emitted for :check-lit IR nodes: we always emit a case
expression.

I also improved how case is compiled, fixing some issues with
ClojureScript. In ClojureScript, it seems, using a list inside list of
cases i.e.

  (case expr
     (form-1 form-2 ,,,) ,,,)

results in a compilation error so we must insure we never compile a
form such as

  (case expr
    ((a b c)) ,,,)

and instead compile

  (case expr
    ([a b c]) ,,,)

Considering this distinction is pretty important, and the current
compilation bug needs to be patched now, I added a dynamic variable
*env* in meander.match.ir.delta for the current &env environment. This
 is bound in meander.match.ir.delta/compile and passed in by
 callers. To be sure, I don't really like this solution as I don't
 like the cognitive overhead of a dynamic variable.