Add the n-times strategy
Adding a combinator I've found particularly useful. There are times where I want to rewrite things and accidentally cause an infinite loop. This strategy allows me to constrain the repetition to some known upper bound. But even more than that, this is very useful when making things that intentionally don't terminate. Examples are things like `l-systems`. I think it will be even more useful when we create non-deterministic rewrite rules.

I didn't add this to readme yet, because we don't really have a great place for it. It might be a good idea to start thinking about splitting out docs into a more structured area than just the readme so we can describe all of these things.