Adds seqable operator
`seqable` initially grew out of my goal to git rid of the comments on scan and vscan about a useless `seq?` check. scan and vscan coerced things using seq, so really they operate on any seqable.

This is where things started and technically it could stop there. I could have made this a purely internal change. But as I thought more about it, I thought this might be a useful operator.

First, it is has already been implicitly asked for. In slack the question of why meander matches on collection type was asked. I personally agree with the justification given, I think this is the right default for meander. At the same time, meander is made to fit into existing code and often that code will produce various seqable collections and it might be nice to express this fact.

`seqable` can also help us to make these things consistent. We could match with `(seqable !xs ...)` and rewrite into `[!xs ...]`.

I did this work partially to understand the various parts of meander and to see what it would take to add a more significant feature like this. I have tried to test things out and ensure they work, but there could be some corner cases.

I'd definitely love to get some feedback on the idea and/or the implementation of it.