Adds simple type information and uses it to remove unncessary checks.
With this change all sorts of redundant checks that we used to do are now removed. Some branches are replaced with failures. I think we can further remove those failing branches, but :match intentionally does not (not sure why). This was motivated by my work on seqable. Those optimizations are no longer needed with this (scan does not have a redundant check). I plan on following up with a defsyntax for seqable.