Define special form meander.epsilon/apply
This patch adds a primitive form of the `app` pattern matching
operator and defines `app` in terms of it, and defines `app` for
substitution also in terms of the same primitive form.

The primitive form of function application is

  (meander.epsilon/apply function argument)

Matching semantics of this form are the same as `app`s with the same
number of arguments.

For substitution this applies the function to the argument.