Rearchitect the expand and parse
This massive patch reworks the design of the expansion and parsing
steps, and migrates all pattern matching and substitution operators in
to the "core" namespace.

There is now only one defsyntax, located in the primary syntax
namespace. The "kind" or "phase" of parsing is now passed as part of
an attribute of the parsing environment. This attribute can be
accessed in the body of defsyntax implementations by looking up
:meander.syntax.epsilon/phase in the &env. This allows syntax
extensions to control the expansion process with respect to context
without the need for multiple defsyntax macros.

All of the pattern matching and substitution operators defined by
meander are now located in the "core" namespace.