Fixes some disjuncts that would fail
When trying to capture in an `or` with multiple literals and a lvr i.e. `(or nil 0 !xs)`, you would get a fail if anything other than a nil or a 0 were passed. This happened because of some flipped around conditionals. This fixes that problem. As far as I can tell it doesn't cause any adverse effects.