Avoid expensive k-combinations with map patterns containing non-literal keys
This patch rewrites map patterns which contain non-literal keys to
avoid the expensive of k-combinations.

A pattern like

    {:a _ :b _ ?c _ ?d _}

is rewritten as

    {:a _ :b _ & {?c _ & {?d _}}}

When the matrix compiler encounters this pattern it will use `seq` to
produce the search space for the non-literal keys as opposed to
`k-combinations` which the first pattern causes.