Avoid expensive k-combinations with set patterns contianing non-literal patterns
This patch rewrites map patterns which contain non-literal keys to
avoid the expensive of k-combinations.

A pattern like

    #{:a :b ?c ?d}

is rewritten as

    #{:a :b ^& #{?c ^& #{?d}}}

When the matrix compiler encounters this pattern it will use `seq` to
produce the search space for the non-literal keys as opposed to
`k-combinations` which the first pattern causes.