Improves performance for nth and repeats of width 1.
For matches like this one:

```
(m/match coll
    [(m/pred number? !xs) ...]
    !xs)
```

Reduces run time to 50%. If we used transients, we could get this to be performance parity with handwritten clojure.

`nth` was improved only for vectors, but I also added type information for :star nodes. Probably could do the same for plus nodes.