meander.epsilon/cata may now be used on the right side of rewrite clauses
This patch ensures `meander.epsilon/cata` may be used on the right
side of `meander.epsilon/rewrite` clauses. This functionality is not
yet supported for `meander.epsilon/rewrites` clauses and an attempt
to use `meander.epsilon/cata` on the right side of
`meander.epsilon/rewrites` clauses will result in a macroexpansion error.