Ignore compiled code. Fix typo. Add local emacs variables.
Anytime I'm in the dev namespaces, I want the dev profile on for deps.edn. This allows cider to use that profile automatically when you jack-in. Since it is not part of the compiled code I figured it should be fine. But we can always remove it later.

I also went ahead and ignored the compiled code. We *might* want to check it in later, but definitely not now.

Finally I fixed where appears to be a typo and was causing my reader to throw and error.