Update pattern generation
This patch incrementally improves pattern generators. Generators now
have the signature

    (-generate gen env) -> [val env* gen*]

where `env*` is the (potentially) modified environment, and `gen*` is
the "next" or "child" node in the generator tree.

Generators can be converted in to streams of values by using the
`run-gen` function in the runtime namespace. For a potentially
infinite number of results use

    (run-gen gen env)

and for a maximum number of results use

    (run-gen gen env limit)

There is still a lot of work to do as some generators produce
incorrect answers. For example

    [!xs ... | . !xs ...]

with the environment

    {!xs [A B]}

will produce in its stream the values

    [A B | B]

and

    [A | A B]

which is wrong!