Begin new approach to building query-function and yield-function
The new approach to building a query functions and yield functions
will deprecate the previous approach. Functions returned by either of
these methods now require three arguments: pass, fail, and state. Pass
and fail are correspond to `:pass` and `:fail` as provided by the
kernel (previously known as runtime/enviornment) used to construct the
function. Controlling the use of pass and fail in the body of the
returned function eliminates the need for the kernel to provide the
`:bind` function.

The motivation for exploring a new approach was encouraged by the poor
performance of code generation. It remains to be seen if this approach
will result in superior performance, however, I surmise that
eliminating the numerous calls to the `:bind` function and, by
extension, calls to pass will improve overall peformance. Each of
these calls either produces, manipulates, or passes through an IR
node. Subsequently, there is IR post-processing which applies rewrite
rules for these nodes. So, hopefully, minimizing or, in the case of
bind, eliminating extra work will improve compiler (and interpreter)
performance.